Add gen-server to xrpc-cli
